+++ packages/ai/core/generate-text/index.ts
@@ -5,6 +5,7 @@ export type { StepResult } from './step-result';
 export { streamText } from './stream-text';
 export type { StreamTextResult, TextStreamPart } from './stream-text-result';
 export type { ToolCallRepairFunction } from './tool-call-repair';
+export { smoothStream } from './smooth-stream';
 
 // TODO 4.1: rename to ToolCall and ToolResult, deprecate old names
 export type {
